Drop dependency on bash - should now run on any posix shell
Replace calls to `echo -e` with `printf`. Flags for echo are undefined
in posix sh.

Replace `==` in `test` commands with `=` since the former is undefined
in posix sh.

Replace `tput cup 0 0` which seems to not work on FreeBSD with `tput
clear`.

Finally, replace `read -s -t 0.5 -n1 key` with stty options and a normal
call to `read`. The the `-s` option is undefined in posix sh and even
for bash, the fractional timeout varies by version.