Restore solaris compatiblity
We are now using /bin/sh to run so to be compatible with Solaris 10,
we cannot use $().

Additionally the builtin read in that shell does not support the -r
option.

The bats test suite is now passing on Solaris (bats itself needs a patch
from https://github.com/sstephenson/bats/pull/79 to work on solaris)